robustness and compatibility fixes
The form [~,keepthis] = funcname(arg1,arg2,...) is not backwards compatible
with older forms of Matlab.
A system call to 'cat' is not compatible with Windows, swapping in a pure
Matlab version.
Also, switching to fullfile and filesep instead of constructing paths
using bare '/' is preferred, as it is OS independent.