robust file location
The location of the www and temp folders were previously implicitly based
off of pwd(), which depends on the directory from which Matlab is invoked.
This has been changed to be relative to the location of the file webserver.m

Moved the test.m function to www/demo_func.m and updated test.py locally to
reflect the changes.

Added 'modern' argument parsing and validation to webserver.m